[YS-408] refactor:  지역 필터링 컴포넌트 분리
